feat:回复功能第二次提交，未完成

+++ src/main/java/com/lhh/community/entity/Notification.java
@@ -16,9 +16,9 @@ public class Notification {
     private Integer id;
 
     /**
-     *
+     * 要通知的人
      */
-    private Long notifier;
+    private Integer notifier;
 
     /**
      *
